fix elapsed time with milliseconds

+++ src/main/java/barcodereader/DecodingResult.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public class DecodingResult {
   public List<BarcodeResult> results;
-  public long elapsedTime;
+  public double elapsedTime;
 }
